requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
docopt==0.6.2
greenlet==2.0.2
idna==3.4
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pipreqs==0.4.11
plotly==5.13.1
protobuf==3.20.3
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tenacity==8.2.2
typing_extensions==4.5.0
urllib3==1.26.15
yarg==0.1.9
Flask~=2.2.3